+++ src/components/MaterialDashboardPro/PictureUpload.js
